requirements.txt
Pillow>=8.0.0
numpy>=1.19.0
wand
bpy
